Split scrap-heavy graph refresh tests
